@@ -22,17 +22,16 @@
     "@types/jquery": "^3.5.6",
     "@types/lodash": "^4.14.168",
     "@types/node": "^12.0.0",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router": "^5.1.12",
-    "@types/react-router-dom": "^5.1.7",
+    "@types/react": "18.0.9",
+    "@types/react-dom": "18.0.4",
+    "@types/react-router-dom": "5.3.3",
     "@types/react-scroll": "^1.8.2",
     "@types/react-transition-group": "^4.4.1",
-    "@types/styled-components": "^5.1.7",
+    "@types/styled-components": "5.1.25",
     "@types/uuid": "8.3.1",
     "@types/video.js": "^7.3.15",
     "@xstate/react": "^1.5.1",
-    "antd": "4.16.13",
+    "antd": "4.20.4",
     "babel-eslint": "^10.1.0",
     "babel-plugin-import": "^1.13.3",
     "craco-less": "^1.17.1",
@@ -51,18 +50,17 @@
     "jquery": "^3.6.0",
     "lodash": "^4.17.21",
     "opensheetmusicdisplay": "1.5.0",
-    "react": "^17.0.1",
+    "react": "18.1.0",
     "react-async-hook": "^4.0.0",
-    "react-dom": "^17.0.1",
+    "react-dom": "18.1.0",
     "react-is": "^17.0.1",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "6.3.0",
     "react-scripts": "4.0.3",
     "react-scroll": "^1.8.1",
     "react-snap": "^1.23.0",
     "react-split-pane": "^0.1.92",
     "resize-observer-polyfill": "^1.5.1",
-    "styled-components": "^5.2.1",
+    "styled-components": "5.3.5",
     "ts-toolbelt": "^9.6.0",
     "typed-graphqlify": "^3.1.4",
     "typescript": "4.4-rc",
