@@ -2,7 +2,7 @@ import { ConfigProvider } from 'antd';
 import enUS from 'antd/lib/locale-provider/en_US';
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
-import { ThemeProvider } from 'styled-components';
+import { ThemeProvider as ThemeProvider_ } from 'styled-components';
 import './App.less';
 import { NewVersionNotifier } from './components/NewVersionNotifier';
 import { Routes } from './components/Routes';
@@ -12,6 +12,8 @@ import { MetaProvider } from './ctx/meta';
 import { RouteInfoProvider } from './ctx/route-info';
 import { ViewportProvider } from './ctx/viewport';
 import { theme } from './theme';
+
+const ThemeProvider = ThemeProvider_ as any;
 
 export const App: React.FC = (props) => {
   return (
