@@ -1,3 +1,4 @@
+import { PropsWithChildren } from 'react';
 import styled from 'styled-components';
 import { useViewport } from '../../ctx/viewport/useViewport';
 
@@ -5,7 +6,7 @@ const Outer = styled.div<{ $height: number }>`
   height: ${(props) => props.$height}px;
 `;
 
-type Props = {};
+type Props = PropsWithChildren<{}>;
 
 export const FullHeightDiv: React.FC<Props> = (props) => {
   const { innerHeight } = useViewport();
