@@ -1,3 +1,5 @@
-import React from 'react';
+import React, { PropsWithChildren } from 'react';
 
-export const NoneLayout: React.FC = (props) => <div data-testid="none-layout">{props.children}</div>;
+export const NoneLayout: React.FC<PropsWithChildren<{}>> = (props) => {
+  return <div data-testid="none-layout">{props.children}</div>;
+};
