@@ -1,12 +1,14 @@
 import { ConfigProvider } from 'antd';
 import enUS from 'antd/lib/locale-provider/en_US';
-import React from 'react';
+import React, { PropsWithChildren } from 'react';
 import { BrowserRouter } from 'react-router-dom';
-import { ThemeProvider } from 'styled-components';
+import { ThemeProvider as ThemeProvider_ } from 'styled-components';
 import { ViewportProvider } from '../ctx/viewport';
 import { theme } from '../theme';
 
-export const Test: React.FC = (props) => {
+const ThemeProvider = ThemeProvider_ as any;
+
+export const Test: React.FC<PropsWithChildren<{}>> = (props) => {
   return (
     <ConfigProvider locale={enUS}>
       <ThemeProvider theme={theme}>
